Report 관련 기능 API Test 완료

+++ src/main/java/com/tenco/jobpotal/report/ReportService.java
@@ -1,5 +1,6 @@
 package com.tenco.jobpotal.report;
 
+import com.tenco.jobpotal._core.errors.exception.Exception403;
 import com.tenco.jobpotal._core.errors.exception.Exception404;
 import com.tenco.jobpotal._core.errors.exception.Exception500;
 import com.tenco.jobpotal.community.userCommunity.UserCommunity;
@@ -48,7 +49,10 @@ public class ReportService {
 
 	// Admin Report List Find All
 	@Transactional
-	public List<ReportResponse.FindAllDTO> findAll() {
+	public List<ReportResponse.FindAllDTO> findAll(LoginUser loginUser) {
+		if (!loginUser.isAdmin()) {
+			throw new Exception403("관리자만 가능한 기능입니다.");
+		}
 		// DB 모든 신고 데이터 조회
 		List<Report> reportList = reportJpaRepository.findAll();
 		// 조회한 Report Entity List를 ReportResponse.FindAllDTO로 변환하여 리턴
@@ -58,7 +62,10 @@ public class ReportService {
 	}
 
 	@Transactional
-	public ReportResponse.DetailDTO findById(Long reportId) {
+	public ReportResponse.DetailDTO findById(Long reportId, LoginUser loginUser) {
+		if (!loginUser.isAdmin()) {
+			throw new Exception403("관리자만 가능한 기능입니다.");
+		}
 		// reportId 받아서 해당 신고 데이터 조회, 없으면 예외처리
 		Report report = reportJpaRepository.findById(reportId)
 				.orElseThrow(() -> new Exception404("해당 신고 내역을 찾을수 없습니다"));
@@ -67,7 +74,11 @@ public class ReportService {
 	}
 
 	@Transactional
-	public void deleteReport(Long reportId) {
+	public void deleteReport(Long reportId, LoginUser loginUser) {
+		// admin login 인지 확인
+		if (!loginUser.isAdmin()) {
+			throw new Exception403("관리자만 가능한 기능입니다.");
+		}
 		Report report = reportJpaRepository.findById(reportId)
 				.orElseThrow(() -> new Exception404("해당 신고 내역을 찾을수 없습니다"));
 		reportJpaRepository.delete(report);
